ref: isolate openrouter chat payload handling
